Convert BigInteger to Long

+++ server/src/main/java/hr/tvz/napredna/java/dijezetserver/controller/LineController.java
@@ -16,7 +16,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -38,12 +37,12 @@ public class LineController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<LineDto> updateLine(@PathVariable BigInteger id, @RequestBody LineDto lineDto) {
+    public ResponseEntity<LineDto> updateLine(@PathVariable Long id, @RequestBody LineDto lineDto) {
         return new ResponseEntity<>(lineService.update(id, lineDto), HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteLine(@PathVariable BigInteger id) {
+    public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         try {
             lineService.deleteById(id);
             return ResponseEntity.noContent().build();
